+++ src/main/java/com/cognite/client/servicesV1/exception/MissingParameterExcetion.java
@@ -1,7 +1,0 @@
-package com.cognite.client.servicesV1.exception;
-
-public class MissingParameterExcetion extends RuntimeException {
-    public MissingParameterExcetion(String errorMessage) {
-        super(errorMessage);
-    }
-}